+++ src/config/dictionary/generics/sources.mjs
@@ -1,6 +1,9 @@
 export const SOURCE_CATEGORIES = {
   legacy: [23, 26],
   excluded: [
-    3, 8, 9, 11, 13, 28,
+    3, 6, 9, 11, 13, 28,
+  ],
+  hidden: [
+    16, 9999999,
   ],
 };